refactor(hush-cli): extract trusted_publishers from pkg_cli
Move cmd_pkg_trusted_publishers and the cmd_trusted_publisher_* handlers
(add/list/remove) into pkg_cli/trusted_publishers.rs, carrying the
#[allow(clippy::too_many_arguments)] on add verbatim.
cmd_pkg_trusted_publishers widened to pub(super); the rest stay private.
add_trusted_publisher_signed_payload dropped from mod.rs's auth import
(its only consumer moved here).
